add like and comment api implementation

+++ service/api/posts-api.go
@@ -9,6 +9,50 @@ import (
 	"github.com/julienschmidt/httprouter"
 )
 
+func (rt *_router) checkUserAccesAndPostAuthor(userId string, requestedUsername string, requestedPostId string) (data string, httpResponse int, err error) {
+	usernameId, err := rt.db.GetUsernameFromId(userId)
+
+	if err != nil {
+		if err.Error() != components.BadRequestError {
+			return components.InternalServerError, http.StatusInternalServerError, err
+		} else {
+			return components.BadRequestError, http.StatusBadRequest, nil
+		}
+	}
+
+	_, err = rt.db.GetIdFromUsername(requestedUsername)
+
+	if err != nil {
+		if err.Error() != components.BadRequestError {
+			return components.InternalServerError, http.StatusInternalServerError, err
+		} else {
+			return components.BadRequestError, http.StatusBadRequest, nil
+		}
+	}
+
+	isBanned, err := rt.db.IsBanished(requestedUsername, usernameId)
+
+	if err != nil {
+		return components.InternalServerError, http.StatusInternalServerError, err
+	}
+
+	if isBanned {
+		return components.ForbiddenError, http.StatusForbidden, nil
+	}
+
+	isAuthor, err := rt.db.ValidPostAuthor(requestedPostId, requestedUsername)
+
+	if err != nil {
+		return components.InternalServerError, http.StatusInternalServerError, err
+	}
+
+	if !isAuthor {
+		return components.BadRequestError, http.StatusBadRequest, nil
+	}
+
+	return "", -1, nil
+}
+
 func (rt *_router) uploadPhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
 	id := r.Header.Get("Authorization")
 
@@ -26,16 +70,16 @@ func (rt *_router) uploadPhoto(w http.ResponseWriter, r *http.Request, ps httpro
 		return
 	}
 
-	var requestBody components.Photo
+	var photo components.Photo
 
-	err = json.NewDecoder(r.Body).Decode(&requestBody)
+	err = json.NewDecoder(r.Body).Decode(&photo)
 
 	if err != nil {
 		HandleResponse(w, ctx, err, "error deconding the request bidy", components.InternalServerError, http.StatusInternalServerError)
 		return
 	}
 
-	data, err := rt.db.UploadPhoto(&requestBody, id)
+	data, err := rt.db.UploadPhoto(&photo, id)
 
 	if err != nil {
 		HandleResponse(w, ctx, err, "error inserting new Photo", data, http.StatusInternalServerError)
@@ -48,7 +92,7 @@ func (rt *_router) uploadPhoto(w http.ResponseWriter, r *http.Request, ps httpro
 func (rt *_router) getUserPosts(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
 	id := r.Header.Get("Authorization")
 
-	_, err := rt.db.GetUsernameFromId(id)
+	usernameId, err := rt.db.GetUsernameFromId(id)
 
 	if err != nil {
 		HandleResponse(w, ctx, err, "error retrieving username of session owner", components.InternalServerError, http.StatusInternalServerError)
@@ -76,10 +120,10 @@ func (rt *_router) getUserPosts(w http.ResponseWriter, r *http.Request, ps httpr
 		return
 	}
 
-	isBannished, err := rt.db.IsBanished(idPosts, id)
+	isBannished, err := rt.db.IsBanished(usernamePosts, usernameId)
 
 	if err != nil {
-		HandleResponse(w, ctx, err, "error trying to search if rewuest owner is banned", components.InternalServerError, http.StatusInternalServerError)
+		HandleResponse(w, ctx, err, "error trying to search if request owner is banned", components.InternalServerError, http.StatusInternalServerError)
 		return
 	}
 
@@ -101,37 +145,23 @@ func (rt *_router) getUserPosts(w http.ResponseWriter, r *http.Request, ps httpr
 func (rt *_router) getPost(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
 	id := r.Header.Get("Authorization")
 
-	_, err := rt.db.GetUsernameFromId(id)
-
-	if err != nil {
-		HandleResponse(w, ctx, err, "error retrieving username of session owner", components.InternalServerError, http.StatusInternalServerError)
-		return
-	}
-
 	usernameReq := ps.ByName("username")
 
-	idReq, err := rt.db.GetIdFromUsername(usernameReq)
+	postId := ps.ByName("post_id")
+
+	response, httpResponse, err := rt.checkUserAccesAndPostAuthor(id, usernameReq, postId)
 
 	if err != nil {
-		HandleResponse(w, ctx, err, "error retriving id of username URL", components.InternalServerError, http.StatusInternalServerError)
+		HandleResponse(w, ctx, err, "error managing Url and session permissions", response, httpResponse)
 		return
 	}
 
-	isBanned, err := rt.db.IsBanished(idReq, id)
-
-	if err != nil {
-		HandleResponse(w, ctx, err, "error checking if the session owner is banned", components.InternalServerError, http.StatusInternalServerError)
+	if httpResponse != -1 {
+		HandleResponse(w, ctx, nil, "", response, httpResponse)
 		return
 	}
 
-	if isBanned {
-		HandleResponse(w, ctx, nil, "", components.ForbiddenError, http.StatusForbidden)
-		return
-	}
-
-	post_id := ps.ByName("post_id")
-
-	data, err := rt.db.GetPost(post_id, id)
+	data, err := rt.db.GetPost(postId, id)
 
 	if err != nil {
 		HandleResponse(w, ctx, err, "error trying to retrieve the post", data, http.StatusInternalServerError)
@@ -161,6 +191,18 @@ func (rt *_router) deletePhoto(w http.ResponseWriter, r *http.Request, ps httpro
 
 	postId := ps.ByName("post_id")
 
+	validAuthor, err := rt.db.ValidPostAuthor(postId, username)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error checking author of the post", components.InternalServerError, http.StatusInternalServerError)
+		return
+	}
+
+	if !validAuthor {
+		HandleResponse(w, ctx, nil, "", components.BadRequestError, http.StatusBadRequest)
+		return
+	}
+
 	data, err := rt.db.DeletePhoto(postId)
 
 	if err != nil {
@@ -169,4 +211,214 @@ func (rt *_router) deletePhoto(w http.ResponseWriter, r *http.Request, ps httpro
 	}
 
 	HandleResponse(w, ctx, nil, "", data, http.StatusNoContent)
+}
+
+func (rt *_router) getLikes(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
+	id := r.Header.Get("Authorization")
+
+	usernameReq := ps.ByName("username")
+
+	postId := ps.ByName("post_id")
+
+	response, httpResponse, err := rt.checkUserAccesAndPostAuthor(id, usernameReq, postId)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error managing Url and session permissions", response, httpResponse)
+		return
+	}
+
+	if httpResponse != -1 {
+		HandleResponse(w, ctx, nil, "", response, httpResponse)
+		return
+	}
+
+	data, err := rt.db.GetLikes(postId, id)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error retriving likers list", data, http.StatusInternalServerError)
+		return
+	}
+
+	HandleResponse(w, ctx, nil, "", data, http.StatusOK)
+}
+
+func (rt *_router) likePhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
+	id := r.Header.Get("Authorization")
+
+	usernameReq := ps.ByName("username")
+
+	postId := ps.ByName("post_id")
+
+	response, httpResponse, err := rt.checkUserAccesAndPostAuthor(id, usernameReq, postId)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error managing Url and session permissions", response, httpResponse)
+		return
+	}
+
+	if httpResponse != -1 {
+		HandleResponse(w, ctx, nil, "", response, httpResponse)
+		return
+	}
+
+	likerId := ps.ByName("liker_id")
+
+	if likerId != id {
+		HandleResponse(w, ctx, nil, "", components.BadRequestError, http.StatusBadRequest)
+		return
+	}
+
+	data, err := rt.db.LikePhoto(postId, likerId)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error inserting the new like", data, http.StatusInternalServerError)
+		return
+	}
+
+	HandleResponse(w, ctx, nil, "", data, http.StatusOK)
+}
+
+func (rt *_router) unlikePhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
+	id := r.Header.Get("Authorization")
+
+	usernameReq := ps.ByName("username")
+
+	postId := ps.ByName("post_id")
+
+	response, httpResponse, err := rt.checkUserAccesAndPostAuthor(id, usernameReq, postId)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error managing Url and session permissions", response, httpResponse)
+		return
+	}
+
+	if httpResponse != -1 {
+		HandleResponse(w, ctx, nil, "", response, httpResponse)
+		return
+	}
+
+	likerId := ps.ByName("liker_id")
+
+	if likerId != id {
+		HandleResponse(w, ctx, nil, "", components.BadRequestError, http.StatusBadRequest)
+		return
+	}
+
+	data, err := rt.db.LikePhoto(postId, likerId)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error inserting the new like", data, http.StatusInternalServerError)
+		return
+	}
+
+	w.WriteHeader(http.StatusNoContent)
+}
+
+func (rt *_router) getComments(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
+	id := r.Header.Get("Authorization")
+
+	usernameReq := ps.ByName("username")
+
+	postId := ps.ByName("post_id")
+
+	response, httpResponse, err := rt.checkUserAccesAndPostAuthor(id, usernameReq, postId)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error managing Url and session permissions", response, httpResponse)
+		return
+	}
+
+	if httpResponse != -1 {
+		HandleResponse(w, ctx, nil, "", response, httpResponse)
+		return
+	}
+
+	data, err := rt.db.GetComments(postId, id)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error retrieving the comments", data, http.StatusInternalServerError)
+		return
+	}
+
+	HandleResponse(w, ctx, nil, "", data, http.StatusOK)
+}
+
+func (rt *_router) commentPhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
+	id := r.Header.Get("Authorization")
+
+	usernameReq := ps.ByName("username")
+
+	postId := ps.ByName("post_id")
+
+	response, httpResponse, err := rt.checkUserAccesAndPostAuthor(id, usernameReq, postId)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error managing Url and session permissions", response, httpResponse)
+		return
+	}
+
+	if httpResponse != -1 {
+		HandleResponse(w, ctx, nil, "", response, httpResponse)
+		return
+	}
+
+	var comment string
+
+	err = json.NewDecoder(r.Body).Decode(&comment)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error retriving body content", components.InternalServerError, http.StatusInternalServerError)
+		return
+	}
+
+	data, err := rt.db.CommentPhoto(postId, id, comment)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error creating comment", data, http.StatusInternalServerError)
+		return
+	}
+	HandleResponse(w, ctx, nil, "", data, http.StatusCreated)
+}
+
+func (rt *_router) uncommentPhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
+	id := r.Header.Get("Authorization")
+
+	usernameReq := ps.ByName("username")
+
+	postId := ps.ByName("post_id")
+
+	response, httpResponse, err := rt.checkUserAccesAndPostAuthor(id, usernameReq, postId)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error managing Url and session permissions", response, httpResponse)
+		return
+	}
+
+	if httpResponse != -1 {
+		HandleResponse(w, ctx, nil, "", response, httpResponse)
+		return
+	}
+
+	isValid, err := rt.db.IsValid(id, usernameReq)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error checking validity of username and Auth", components.InternalServerError, http.StatusInternalServerError)
+		return
+	}
+
+	if !isValid {
+		HandleResponse(w, ctx, nil, "", components.UnauthorizedError, http.StatusUnauthorized)
+		return
+	}
+
+	commentId := ps.ByName("comment_id")
+
+	data, err := rt.db.UncommentPhoto(commentId)
+
+	if err != nil {
+		HandleResponse(w, ctx, err, "error deleting the comment", data, http.StatusInternalServerError)
+		return
+	}
+
+	w.WriteHeader(http.StatusOK)
 }
